core compatiblity

+++ sfs-main-extension/src/com/gerantech/towers/sfs/battle/BattleRoom.java
@@ -147,7 +147,7 @@ public class BattleRoom extends SFSExtension
 
 				checkBattleEnding(battleDuration);
 			}
-		}, 0, (int) battleField.interval, TimeUnit.MILLISECONDS);
+		}, 0, battleField.interval, TimeUnit.MILLISECONDS);
 
 		trace(room.getName(), "created.");
 	}
